Split verify/saga
