+++ packages/mobile/src/app/saga.ts
@@ -38,8 +38,7 @@ import { handlePaymentDeeplink } from 'src/send/utils'
 import { navigateToURI } from 'src/utils/linking'
 import Logger from 'src/utils/Logger'
 import { clockInSync } from 'src/utils/time'
-import { receiveAttestationCode } from 'src/verify/reducer'
-import { CodeInputType } from 'src/verify/saga'
+import { CodeInputType, receiveAttestationCode } from 'src/verify/module'
 import { parse } from 'url'
 
 const TAG = 'app/saga'
