+++ packages/mobile/src/identity/privateHashing.test.ts
@@ -10,7 +10,7 @@ import { ErrorMessages } from 'src/app/ErrorMessages'
 import { updateE164PhoneNumberSalts } from 'src/identity/actions'
 import { fetchPhoneHashPrivate } from 'src/identity/privateHashing'
 import { e164NumberToSaltSelector } from 'src/identity/reducer'
-import { isBalanceSufficientForSigRetrievalSelector } from 'src/verify/reducer'
+import { isBalanceSufficientForSigRetrievalSelector } from 'src/verify/module'
 import { isAccountUpToDate } from 'src/web3/dataEncryptionKey'
 import { getConnectedAccount } from 'src/web3/saga'
 import { createMockStore } from 'test/utils'
