+++ packages/mobile/src/send/Send.test.tsx
@@ -3,7 +3,7 @@ import { render } from 'react-native-testing-library'
 import { Provider } from 'react-redux'
 import { Screens } from 'src/navigator/Screens'
 import Send from 'src/send/Send'
-import { KomenciAvailable } from 'src/verify/reducer'
+import { KomenciAvailable } from 'src/verify/module'
 import { createMockStore, getMockStackScreenProps } from 'test/utils'
 import {
   mockE164Number,
