+++ packages/mobile/src/verify/saga.test.ts
@@ -1,13 +1,10 @@
 import { ActionableAttestation } from '@celo/contractkit/lib/wrappers/Attestations'
-import { verifyWallet } from '@celo/komencikit/src/verifyWallet'
 import { getPhoneHash } from '@celo/utils/lib/phoneNumbers'
 import * as reduxSagaTestPlan from 'redux-saga-test-plan'
 import { throwError } from 'redux-saga-test-plan/providers'
 import { call, delay, select } from 'redux-saga/effects'
-import { VerificationEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { ErrorMessages } from 'src/app/ErrorMessages'
-import networkConfig from 'src/geth/networkConfig'
 import { celoTokenBalanceSelector } from 'src/goldToken/selectors'
 import { updateE164PhoneNumberSalts } from 'src/identity/actions'
 import { KomenciErrorQuotaExceeded } from 'src/identity/feelessVerificationErrors'
@@ -17,7 +14,10 @@ import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { waitFor } from 'src/redux/sagas-helpers'
 import { stableTokenBalanceSelector } from 'src/stableToken/reducer'
+import { fetchKomenciSession, getKomenciKit } from 'src/verify/komenci'
 import {
+  AttestationCode,
+  BALANCE_CHECK_TIMEOUT,
   e164NumberSelector,
   ensureRealHumanUser,
   fail,
@@ -28,38 +28,25 @@ import {
   KomenciAvailable,
   komenciContextSelector,
   phoneHashSelector,
-  reportRevealStatus,
   requestAttestations,
   setActionableAttestation,
   setKomenciAvailable,
-  setKomenciContext,
   setPhoneHash,
   setVerificationStatus,
   shouldUseKomenciSelector,
   start,
-} from 'src/verify/reducer'
+} from 'src/verify/module'
 import {
-  AttestationCode,
-  BALANCE_CHECK_TIMEOUT,
-  checkIfKomenciAvailableSaga,
   failSaga,
-  fetchKomenciReadiness,
-  fetchKomenciSession,
   fetchOnChainDataSaga,
-  fetchOrDeployMtwSaga,
   fetchPhoneNumberDetailsSaga,
   getActionableAttestations,
-  getKomenciKit,
-  getPhoneHashDetails,
-  reportActionableAttestationsStatuses,
-  reportRevealStatusSaga,
   resetSaga,
   startSaga,
 } from 'src/verify/saga'
 import { getContractKit, getContractKitAsync } from 'src/web3/contracts'
-import { registerWalletAndDekViaKomenci } from 'src/web3/dataEncryptionKey'
 import { getAccount, getConnectedUnlockedAccount, unlockAccount, UnlockResult } from 'src/web3/saga'
-import { mockE164NumberHash, mockE164NumberPepper, mockPublicDEK } from 'test/values'
+import { mockE164NumberPepper, mockPublicDEK } from 'test/values'
 
 export const mockAccount = '0x0000000000000000000000000000000000007E57'
 export const mockAccount1 = '0x0000000000000000000000000000000000007E58'
@@ -80,9 +67,9 @@ export const mockPhoneHashDetails = {
   pepper: mockE164NumberPepper,
 }
 
-const MockedAnalytics = ValoraAnalytics as any
+export const MockedAnalytics = ValoraAnalytics as any
 
-const mockActionableAttestation = {
+export const mockActionableAttestation = {
   issuer: mockAccount2,
   blockNumber: 1,
   attestationServiceURL: 'https://example.com',
@@ -90,25 +77,25 @@ const mockActionableAttestation = {
   version: '1.0.0',
 }
 
-const attestationCode0: AttestationCode = {
+export const attestationCode0: AttestationCode = {
   code:
     'ab8049b95ac02e989aae8b61fddc10fe9b3ac3c6aebcd3e68be495570b2d3da15aabc691ab88de69648f988fab653ac943f67404e532cfd1013627f56365f36501',
   issuer: '848920b14154b6508b8d98e7ee8159aa84b579a4',
 }
 
-const attestationCode1: AttestationCode = {
+export const attestationCode1: AttestationCode = {
   code:
     '2033a9e1268576bf5dfee354a37480529d71f99be82c05005ffb71c7d742d10e7a9aa01f8acc4d7998e1e8b183cf6b8cb4d4a8d923fecfddd191e61e074adc5e00',
   issuer: 'fdb8da92c3597e81c2737e8be793bee9f1172045',
 }
 
-const attestationCode2: AttestationCode = {
+export const attestationCode2: AttestationCode = {
   code:
     '1930a9e1268576bf5dfee354a37480529d71f99be82c05005ffb71c7d742d10e7a9aa01f8acc4d7993f75ab183cf6b8cb4d4a8d923fecfddd191e61e074adc5a10',
   issuer: 'ecb8da92c3597e81c2737e8be793bee9f1173156',
 }
 
-const mockActionableAttestations: ActionableAttestation[] = [
+export const mockActionableAttestations: ActionableAttestation[] = [
   {
     issuer: attestationCode0.issuer,
     blockNumber: 100,
@@ -132,40 +119,22 @@ const mockActionableAttestations: ActionableAttestation[] = [
   },
 ]
 
-const mockKomenciKit = {
+export const mockKomenciKit = {
   getDistributedBlindedPepper: jest.fn(),
   deployWallet: jest.fn(),
 }
 
-const mockAccountsWrapper = {
+export const mockAccountsWrapper = {
   getWalletAddress: jest.fn(() => Promise.resolve(mockAccount)),
   getDataEncryptionKey: jest.fn(() => Promise.resolve(mockPublicDEK)),
 }
 
-const mockAttestationsWrapper = {
+export const mockAttestationsWrapper = {
   lookupAccountsForIdentifier: jest.fn(),
   getVerifiedStatus: jest.fn(),
   getRevealStatus: jest.fn(),
   getActionableAttestations: jest.fn(),
 }
-
-describe(checkIfKomenciAvailableSaga, () => {
-  it('sets komenci availability', async () => {
-    const contractKit = await getContractKitAsync()
-    const komenciKit = getKomenciKit(contractKit, mockAccount, mockKomenciContext)
-    await reduxSagaTestPlan
-      .expectSaga(checkIfKomenciAvailableSaga)
-      .provide([
-        [call(getContractKit), contractKit],
-        [call(getAccount), mockAccount],
-        [select(komenciContextSelector), mockKomenciContext],
-        [call(getKomenciKit, contractKit, mockAccount, mockKomenciContext), komenciKit],
-        [call(fetchKomenciReadiness, komenciKit), true],
-      ])
-      .put(setKomenciAvailable(KomenciAvailable.Yes))
-      .run()
-  })
-})
 
 describe(startSaga, () => {
   it('starts with Komenci active session', async () => {
@@ -403,196 +372,6 @@ describe(fetchPhoneNumberDetailsSaga, () => {
   })
 })
 
-describe(fetchOrDeployMtwSaga, () => {
-  it('fails on multiple verified addresses', async () => {
-    const contractKit = await getContractKitAsync()
-    const komenciKit = getKomenciKit(contractKit, mockAccount, mockKomenciContext)
-    ;(mockAttestationsWrapper.lookupAccountsForIdentifier as jest.Mock).mockReturnValue(['0', '1'])
-    ;(mockAttestationsWrapper.getVerifiedStatus as jest.Mock).mockReturnValue({
-      isVerified: true,
-    })
-    await reduxSagaTestPlan
-      .expectSaga(fetchOrDeployMtwSaga)
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [call(getContractKit), contractKit],
-        [call(getConnectedUnlockedAccount), mockAccount],
-        [select(komenciContextSelector), mockKomenciContext],
-        [call(getKomenciKit, contractKit, mockAccount, mockKomenciContext), komenciKit],
-        [select(shouldUseKomenciSelector), false],
-        [select(phoneHashSelector), mockPhoneHash],
-        [
-          call([contractKit.contracts, contractKit.contracts.getAttestations]),
-          mockAttestationsWrapper,
-        ],
-        [
-          call(
-            verifyWallet,
-            contractKit,
-            '0',
-            networkConfig.allowedMtwImplementations,
-            mockAccount
-          ),
-          { ok: true },
-        ],
-        [
-          call(
-            verifyWallet,
-            contractKit,
-            '1',
-            networkConfig.allowedMtwImplementations,
-            mockAccount
-          ),
-          { ok: true },
-        ],
-      ])
-      .put(
-        fail('More than one verified MTW with walletAddress as signer found. Should never happen')
-      )
-      .run()
-  })
-
-  it('succeeds for fresh new account', async () => {
-    const contractKit = await getContractKitAsync()
-    ;(mockAttestationsWrapper.lookupAccountsForIdentifier as jest.Mock).mockReturnValue([])
-    ;(mockAttestationsWrapper.getVerifiedStatus as jest.Mock).mockReturnValue({
-      isVerified: true,
-    })
-    ;(mockKomenciKit.deployWallet as jest.Mock).mockReturnValue({ ok: true, result: mockAccount1 })
-    const mockKomenciContextActive = {
-      ...mockKomenciContext,
-      sessionActive: true,
-    }
-    const mockVerifyWallet = jest.fn()
-    ;(mockVerifyWallet as jest.Mock).mockReturnValueOnce({
-      ok: true,
-    })
-    const mockRegisterWalletAndDekViaKomenci = jest.fn()
-    await reduxSagaTestPlan
-      .expectSaga(fetchOrDeployMtwSaga)
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [call(getContractKit), contractKit],
-        [call(getConnectedUnlockedAccount), mockAccount],
-        [select(komenciContextSelector), mockKomenciContextActive],
-        [call(getKomenciKit, contractKit, mockAccount, mockKomenciContextActive), mockKomenciKit],
-        [select(phoneHashSelector), mockPhoneHash],
-        [
-          call([contractKit.contracts, contractKit.contracts.getAttestations]),
-          mockAttestationsWrapper,
-        ],
-        {
-          call: ({ fn }, next) => (fn === verifyWallet ? mockVerifyWallet() : next()),
-        },
-        {
-          call: ({ fn }, next) =>
-            fn === registerWalletAndDekViaKomenci ? mockRegisterWalletAndDekViaKomenci() : next(),
-        },
-      ])
-      .put(setKomenciContext({ unverifiedMtwAddress: mockAccount1 }))
-      .put(fetchOnChainData())
-      .run()
-    expect(mockRegisterWalletAndDekViaKomenci.mock.calls.length).toBe(1)
-  })
-
-  it('succeeds for already cached unverified MTW address', async () => {
-    const contractKit = await getContractKitAsync()
-    const komenciKit = getKomenciKit(contractKit, mockAccount, mockKomenciContext)
-    ;(mockAttestationsWrapper.lookupAccountsForIdentifier as jest.Mock).mockReturnValue([
-      mockAccount1,
-    ])
-    const mockKomenciContextWithUnverifiedMtwAddress = {
-      ...mockKomenciContext,
-      unverifiedMtwAddress: mockAccount1,
-      sessionActive: true,
-    }
-    const mockVerifyWallet = jest.fn()
-    ;(mockVerifyWallet as jest.Mock).mockReturnValueOnce({
-      ok: false,
-    })
-    ;(mockVerifyWallet as jest.Mock).mockReturnValueOnce({
-      ok: true,
-    })
-    const mockRegisterWalletAndDekViaKomenci = jest.fn()
-    await reduxSagaTestPlan
-      .expectSaga(fetchOrDeployMtwSaga)
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [call(getContractKit), contractKit],
-        [call(getConnectedUnlockedAccount), mockAccount],
-        [select(komenciContextSelector), mockKomenciContextWithUnverifiedMtwAddress],
-        [
-          call(getKomenciKit, contractKit, mockAccount, mockKomenciContextWithUnverifiedMtwAddress),
-          komenciKit,
-        ],
-        [select(phoneHashSelector), mockPhoneHash],
-        [
-          call([contractKit.contracts, contractKit.contracts.getAttestations]),
-          mockAttestationsWrapper,
-        ],
-        {
-          call: ({ fn }, next) => (fn === verifyWallet ? mockVerifyWallet() : next()),
-        },
-        {
-          call: ({ fn }, next) =>
-            fn === registerWalletAndDekViaKomenci ? mockRegisterWalletAndDekViaKomenci() : next(),
-        },
-      ])
-      .put(setKomenciContext({ unverifiedMtwAddress: mockAccount1 }))
-      .put(fetchOnChainData())
-      .run()
-    expect(mockRegisterWalletAndDekViaKomenci.mock.calls.length).toBe(1)
-  })
-
-  it('succeeds if already deployed wallet is a valid MTW', async () => {
-    const contractKit = await getContractKitAsync()
-    const komenciKit = getKomenciKit(contractKit, mockAccount, mockKomenciContext)
-    ;(mockAttestationsWrapper.lookupAccountsForIdentifier as jest.Mock).mockReturnValue([
-      mockAccount1,
-    ])
-    ;(mockAttestationsWrapper.getVerifiedStatus as jest.Mock).mockReturnValue({
-      isVerified: true,
-    })
-
-    await reduxSagaTestPlan
-      .expectSaga(fetchOrDeployMtwSaga)
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [call(getContractKit), contractKit],
-        [call(getConnectedUnlockedAccount), mockAccount],
-        [select(komenciContextSelector), mockKomenciContext],
-        [call(getKomenciKit, contractKit, mockAccount, mockKomenciContext), komenciKit],
-        [select(shouldUseKomenciSelector), false],
-        [select(phoneHashSelector), mockPhoneHash],
-        [
-          call([contractKit.contracts, contractKit.contracts.getAttestations]),
-          mockAttestationsWrapper,
-        ],
-        [
-          call(
-            verifyWallet,
-            contractKit,
-            mockAccount1,
-            networkConfig.allowedMtwImplementations,
-            mockAccount
-          ),
-          { ok: true },
-        ],
-      ])
-      .put(
-        setKomenciContext({
-          unverifiedMtwAddress: mockAccount1,
-        })
-      )
-      .put(
-        setVerificationStatus({
-          isVerified: true,
-        })
-      )
-      .run()
-  })
-})
-
 describe(fetchOnChainDataSaga, () => {
   it('succeeds with Komenci enabled', async () => {
     const contractKit = await getContractKitAsync()
@@ -698,102 +477,6 @@ describe(failSaga, () => {
     await reduxSagaTestPlan
       .expectSaga(failSaga, 'test')
       .provide([[select(shouldUseKomenciSelector), true]])
-      .run()
-  })
-})
-
-describe(reportRevealStatusSaga, () => {
-  beforeEach(() => {
-    MockedAnalytics.track.mockReset()
-  })
-  it('report actionable attestation to analytics', async () => {
-    const contractKit = await getContractKitAsync()
-    ;(mockAttestationsWrapper.getRevealStatus as jest.Mock).mockReturnValue({
-      ok: true,
-      json: () => body,
-    })
-
-    const mockIssuer = mockActionableAttestation.issuer
-    const body = { issuer: mockIssuer, custom: 'payload' }
-    await reduxSagaTestPlan
-      .expectSaga(reportRevealStatusSaga, {
-        payload: {
-          attestationServiceUrl: 'url',
-          e164Number: mockE164Number,
-          account: mockAccount,
-          issuer: mockIssuer,
-          pepper: mockE164NumberPepper,
-        },
-      })
-      .provide([
-        [
-          call([contractKit.contracts, contractKit.contracts.getAttestations]),
-          mockAttestationsWrapper,
-        ],
-      ])
-      .run()
-    expect(MockedAnalytics.track.mock.calls.length).toBe(1)
-    expect(MockedAnalytics.track.mock.calls[0][0]).toBe(
-      VerificationEvents.verification_reveal_attestation_status
-    )
-    expect(MockedAnalytics.track.mock.calls[0][1]).toStrictEqual(body)
-  })
-})
-
-describe(reportActionableAttestationsStatuses, () => {
-  it('report actionable attestations', async () => {
-    const contractKit = await getContractKitAsync()
-    ;(mockAttestationsWrapper.getActionableAttestations as jest.Mock).mockReturnValue(
-      mockActionableAttestations
-    )
-
-    await reduxSagaTestPlan
-      .expectSaga(reportActionableAttestationsStatuses)
-      .provide([
-        [call(getConnectedUnlockedAccount), mockAccount],
-        [call(getPhoneHashDetails), mockPhoneHashDetails],
-        [select(e164NumberSelector), mockE164Number],
-        [
-          call([contractKit.contracts, contractKit.contracts.getAttestations]),
-          mockAttestationsWrapper,
-        ],
-        [call([contractKit.contracts, contractKit.contracts.getAccounts]), mockAccountsWrapper],
-        [
-          call(fetchPhoneHashPrivate, mockE164Number),
-          {
-            phoneHash: mockE164NumberHash,
-            e164Number: mockE164Number,
-            pepper: mockE164NumberPepper,
-          },
-        ],
-      ])
-      .put(
-        reportRevealStatus({
-          attestationServiceUrl: mockActionableAttestations[0].attestationServiceURL,
-          account: mockAccount,
-          issuer: mockActionableAttestations[0].issuer,
-          e164Number: mockE164Number,
-          pepper: mockE164NumberPepper,
-        })
-      )
-      .put(
-        reportRevealStatus({
-          attestationServiceUrl: mockActionableAttestations[1].attestationServiceURL,
-          account: mockAccount,
-          issuer: mockActionableAttestations[1].issuer,
-          e164Number: mockE164Number,
-          pepper: mockE164NumberPepper,
-        })
-      )
-      .put(
-        reportRevealStatus({
-          attestationServiceUrl: mockActionableAttestations[2].attestationServiceURL,
-          account: mockAccount,
-          issuer: mockActionableAttestations[2].issuer,
-          e164Number: mockE164Number,
-          pepper: mockE164NumberPepper,
-        })
-      )
       .run()
   })
 })
