+++ packages/mobile/test/schemas.ts
@@ -3,8 +3,7 @@ import { PincodeType } from 'src/account/reducer'
 import { AppState } from 'src/app/actions'
 import { DEFAULT_DAILY_PAYMENT_LIMIT_CUSD } from 'src/config'
 import { RootState } from 'src/redux/reducers'
-import { idle, KomenciAvailable } from 'src/verify/reducer'
-import { NUM_ATTESTATIONS_REQUIRED } from 'src/verify/saga'
+import { idle, KomenciAvailable, NUM_ATTESTATIONS_REQUIRED } from 'src/verify/module'
 
 // Default (version -1 schema)
 export const vNeg1Schema = {
